cleans up reducer/epic configuration to be closer to the screen
the state tree now represents the individual screens and doesn't (yet) have the concept of shared state across the app. That could be easily included as needed. Currently we aren't considering login/auth so the instructor gets loaded explicitly. IRL we will likely load initially based on who is looking at it.