destructure LessonNextSteps for versatility (#150)
This breaks LessonNextSteps apart so that the status of a lesson is its own component, and next steps is strictly concerned with actions. Also added some colors for buttons and modified some text.

The LessonTeaser needs lots of design love. Was thinking about that here: https://paper.dropbox.com/doc/Lists-of-Lessons-LOL-hkj93JkuRH5ggz45BHrvJ